Cache by node counter, restore reading without DB lookup

+++ cmd/utils/flags.go
@@ -1071,7 +1071,7 @@ func SetEthConfig(ctx *cli.Context, stack *node.Node, cfg *eth.Config) {
 	}
 	// TODO(fjl): move trie cache generations into config
 	if gen := ctx.GlobalInt(TrieCacheGenFlag.Name); gen > 0 {
-		state.MaxTrieCacheGen = uint16(gen)
+		state.MaxTrieCacheGen = uint32(gen)
 	}
 }
