+++ trie/secure_trie.go
@@ -59,7 +59,7 @@ type SecureTrie struct {
 // Loaded nodes are kept around until their 'cache generation' expires.
 // A new cache generation is created by each call to Commit.
 // cachelimit sets the number of past cache generations to keep.
-func NewSecure(root common.Hash, db Database, cachelimit uint16, prefix []byte, blockNr uint32) (*SecureTrie, error) {
+func NewSecure(root common.Hash, db Database, cachelimit uint32, prefix []byte, blockNr uint32) (*SecureTrie, error) {
 	if db == nil {
 		panic("NewSecure called with nil database")
 	}
